+++ CHANGELOG.md
@@ -13,6 +13,9 @@ This project adheres to Semantic Versioning.
 * Color now implements `Add`, `AddAssign`, `Sub`, `SubAssign`, `Mul`, `MulAssign`, `Div` and `DivAssign.`
     * For convenience's sake, these are currently implemented as saturating operations, preventing values from going outside the valid range for a color.
     * In future versions, this clamping behaviour might be applied to the type as a whole, but this would require breaking changes.
+* `Texture` and `Canvas` now have methods for writing pixel data at runtime:
+    * `set_data` writes to a specified region of the texture.
+    * `replace_data` is a shortcut method for replacing the entire texture's data at once.
 
 ### Changed